put max proof in same file
