+++ src/basic_block.rs
@@ -14,8 +14,7 @@ pub use div::{DivConstBasicBlock, DivScalarBasicBlock};
 pub use eq::EqBasicBlock;
 pub use id::IdBasicBlock;
 pub use matmul::MatMulBasicBlock;
-pub use max::MaxBasicBlock;
-pub use max_v1::MaxV1BasicBlock;
+pub use max::{MaxBasicBlock, MaxProofBasicBlock};
 pub use mul::{MulBasicBlock, MulConstBasicBlock, MulScalarBasicBlock};
 use ndarray::{ArrayD, IxDyn};
 pub use ops::*;
@@ -40,7 +39,6 @@ pub mod eq;
 pub mod id;
 pub mod matmul;
 pub mod max;
-pub mod max_v1;
 pub mod mul;
 pub mod ops;
 pub mod permute;
