+++ src/tests.rs
@@ -98,7 +98,7 @@ fn test_max() {
   let srs = &ptau::load_file("challenge", 7, 7);
   let empty = ArrayD::zeros(IxDyn(&[0]));
   let a = ArrayD::from_shape_vec(IxDyn(&[4]), vec![-4, 2, -1, 4].into_iter().map(|x| Fr::from(x)).collect()).unwrap();
-  testBasicBlock(MaxV1BasicBlock {}, srs, &empty, &vec![&a]);
+  testBasicBlock(MaxProofBasicBlock {}, srs, &empty, &vec![&a]);
 }
 
 #[test]
